Update template Structure. Include: Docker-compose, PostgreSQL, RestFramework, admin...
