+++ .vscode/settings.json
@@ -1,3 +1,4 @@
 {
-    "python.pythonPath": "venv/bin/python"
+    "python.pythonPath": "/Users/mba0051/.pyenv/shims/python",
+    "python.formatting.provider": "autopep8"
 }
