--- auth_api/auth_api/urls.py
@@ -1,31 +1,0 @@
-"""auth_api URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-from django.conf.urls import url, include
-from django.contrib import admin
-from rest_framework.documentation import include_docs_urls
-
-# Add view from app.
-from auth_api import views
-
-urlpatterns = [
-    url(r'^admin/', admin.site.urls),
-    url(r'^docs/', include_docs_urls(title='Todo API',
-                                     description='RESTful API for Todo')),
-
-    url(r'^$', views.api_root),
-    url(r'^', include(('users.urls', 'users'), namespace='users')),
-    url(r'^', include(('todos.urls', 'todos'), namespace='todos')),
-]
