--- auth_api/todos/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TodosConfig(AppConfig):
-    name = 'todos'
