--- auth_api/users/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class UsersConfig(AppConfig):
-    name = 'users'
