--- auth_api/users/models.py
@@ -1,9 +1,0 @@
-from django.db import models
-
-# Create your models here.
-from django.db import models
-from django.contrib.auth.models import AbstractUser
- 
- 
-class User(AbstractUser):
-    pass
