--- dbconnect/requirements/base.txt
@@ -1,5 +1,0 @@
-Django==2.1.4
-django-filter==2.0.0
-djangorestframework==3.9.0
-mysqlclient==1.3.14
-pytz==2018.7
