--- dbconnect/templates/todo/index.html
@@ -1,12 +1,0 @@
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>Demo page by Python</title>
-        {% load static %}
-    </head>
-    <body>
-        <h1>Todo list</h1>
-
-        <script src="{% static 'js/todo.js' %}"></script>
-    </body>
-</html>
