+++ ddp/Pipfile
@@ -1,0 +1,18 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+django = "*"
+django-filter = "*"
+coreapi = "==2.3.3"
+djangorestframework = "==3.9.0"
+
+gunicorn = "==19.9.0"
+psycopg2-binary = "*"
+
+[dev-packages]
+
+[requires]
+python_version = "3.7"
