+++ ddp/Procfile
@@ -1,0 +1,1 @@
+web: gunicorn ddp.wsgi --log-file -
