+++ ddp/posts/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import Post
+
+admin.site.register(Post)
