+++ ddp/posts/tests.py
@@ -1,0 +1,2 @@
+from django.test import TestCase
+
