+++ ddp/posts/urls.py
@@ -1,0 +1,8 @@
+from posts import views
+from django.conf.urls import url
+from rest_framework.urlpatterns import format_suffix_patterns
+
+urlpatterns = [
+    #url(r'^posts/$', views.PostList.as_view(), name='post-list'),
+    #url(r'^posts/(?P<pk>[0-9]+)/$', views.PostDetail.as_view(), name='post-detail'),
+]
