+++ ddp/topics/serializers.py
@@ -1,0 +1,14 @@
+from rest_framework import serializers
+
+from .models import Topic
+
+
+class TopicSerializer(serializers.HyperlinkedModelSerializer): 
+    class Meta:
+        model = Topic
+        fields = ('id', 'text')
+        extra_kwargs = {
+            'url': {
+                'view_name': 'topic:topic-detail',
+            }
+        }
